requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.1.3
django-filter==22.1
django-model-utils==4.3.1
djangorestframework==3.14.0
Faker==15.3.4
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
whitenoise==6.2.0
